requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Speech Processing
librosa>=0.10.0
soundfile>=0.12.0
pyaudio>=0.2.11
speech-recognition>=3.10.0
pydub>=0.25.0

# Text Processing and NLP
nltk>=3.8.0
spacy>=3.6.0
textblob>=0.17.0
vaderSentiment>=3.3.0

# Explainable AI
shap>=0.42.0
lime>=0.2.0
captum>=0.6.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Web Interface
streamlit>=1.25.0
gradio>=3.40.0
flask>=2.3.0

# Audio/Video Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Development Tools
jupyter>=1.0.0
ipython>=8.14.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Model Optimization
optuna>=3.2.0
hyperopt>=0.2.7

# GPU Support (optional)
# nvidia-ml-py3>=7.352.0
